Refactor screens into modules
